nix: expose anyrun provider under passthru; assert when unavailable
This is no doubt a hacky solution, but it aims to catch an edge case. It
should be unlikely that the user is using the Anyrun module from the
flake and the Anyrun package from Nixpkgs. Nevertheless, we correctly
inform the user that what they're doing is not supported. If the user
insists on using `pkgs.anyrun`, then they certainly have Home Manager
available and thus they may avoid using the flake altogether.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: Iefaf6915551aad04ea08e14bce16d0fa6a6a6964